+++ controller/booking.py
@@ -1,4 +1,5 @@
-from model.booking import Booking
+from model.booking import BookingModel
+from view.booking import BookingView
 from dotenv import get_key
 import time
 import jwt
@@ -27,16 +28,12 @@ def bookingAttraction():
 
 		data = request.get_json()
 		memberId = userInfo["id"]
-		attractionId = data['attractionId']
-		date = data['date']
-		time = data['time']
-		price = data['price']
+		attractionId, date, time, price = BookingView.attraction(data)
 
 		if not memberId or not attractionId or not date or not time or not price or datetime.strptime(date, '%Y-%m-%d').date() <= datetime.today().date():
 			return error(result,"建立失敗，輸入不正確(例：無法選今天及過去日期)或其他原因"), 400
 
-		booking = Booking()
-		booking.post_booking(memberId, attractionId, date, time, price)
+		BookingModel.post_booking(memberId, attractionId, date, time, price)
 
 		result["ok"] = True
 		return result, 200
@@ -56,27 +53,9 @@ def getBooking():
 
 		memberId = userInfo["id"]
 
-		booking = Booking()
-		data = booking.get_booking(memberId)
+		data = BookingModel.get_booking(memberId)
 
-		result["data"] = []
-
-		if len(data) == 0:
-			result["data"] = None
-			return result, 200
-
-		for i in range(len(data)):
-			item = {}
-			item["attraction"] = {}
-			item["attraction"]["id"] = data[i][0]
-			item["attraction"]["name"] = data[i][1]
-			item["attraction"]["address"] = data[i][2]
-			item["attraction"]["image"] = data[i][3].split(',')[0]
-			item["date"] = data[i][4]
-			item["time"] = data[i][5]
-			item["price"] = data[i][6]
-			item["bookingId"] = data[i][7]
-			result["data"].append(item)
+		result = BookingView.booking(result, data)
 		return result, 200		
 
 	except Exception as e:
@@ -96,8 +75,7 @@ def deleteBooking():
 		bookingId = data["bookingId"]
 		memberId = userInfo["id"]
 
-		booking = Booking()
-		booking.delete_booking(memberId,bookingId)
+		BookingModel.delete_booking(memberId,bookingId)
 
 		result["ok"] = True
 		return result, 200
